requirements-index.txt
# Vector search and indexing dependencies
qdrant-client==1.9.0
sentence-transformers==2.7.0
langchain==0.2.0
langchain-community==0.2.0
langchain-qdrant==0.1.0
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
prometheus-client==0.20.0